+++ packages/tools/utils/src/iterable/array.js
@@ -4,7 +4,7 @@
 
 import * as _ from "lamb";
 
-import {tapValue} from "../any/any";
+import {tapValue} from "../string/fn/any/any";
 
 /**
  * Print the provided array and element and return the result of appending the element to the array
@@ -29,5 +29,5 @@ tapAppendTo("abc", {a: 1})
  */
 export const tapAppendTo = _.tapArgs(
     _.appendTo,
-    [tapValue, tapValue]
+    [tapValue(), tapValue()]
 );