+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 
 - updated docs
 
+### Any -> Any
+
+- `tapValue`, `tapType`, `tapTypeAndValue` now expect an optional string and return a tap function
+
 ### Any -> (Array -> Object)
 
 - added `makeKeyed`